Create separate file for Item's methods
