+++ managers/item_manager.rb
@@ -1,0 +1,8 @@
+def list_items(class_name, items)
+  items.select { |item| item.is_a?(class_name) }
+end
+
+def add_item(item, items)
+  items << item
+  puts "#{item.class.name} added successfully."
+end
